# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Network School OAuth Provider",
    "description": "OAuth 2.0 / OpenID Connect identity provider for Network School.\n\nLets third-party apps request NS user data via standard OAuth scopes \u2014 the same pattern as \"Sign in with Google.\"\n\n## Flows\n\n- **Authorization Code + PKCE** \u2014 user-facing \"Sign in with Network School\" flow\n- **Client Credentials** \u2014 machine-to-machine, no user context\n\n## Scopes\n\n| Scope | Claims |\n|-------|--------|\n| `openid` | `sub` |\n| `email` | `email`, `email_verified` |\n| `profile` | `name`, `picture` |\n| `roles` | `roles` (Discord roles, live) |\n| `date_joined` | `date_joined` |\n| `offline_access` | refresh tokens |\n\n## Authentication\n\n- **User auth**: Discord OAuth2 (NS guild membership required)\n- **Sessions**: HS256 JWT in httponly cookie (`ns_session`)\n- **Access tokens**: RS256 JWT (auto-generated RSA keys)\n- **Client auth**: `client_secret` verified via bcrypt\n",
    "license": {
      "name": "Private",
      "url": "https://ns.com/"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/apps/": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "List all OAuth apps",
        "description": "Returns all registered OAuth applications, sorted by creation date (newest first). Does not include client secrets.",
        "operationId": "list_apps_api_apps__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OAuthAppResponse"
                  },
                  "type": "array",
                  "title": "Response List Apps Api Apps  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apps"
        ],
        "summary": "Register a new OAuth app",
        "description": "Creates a new OAuth client application. Returns the `client_id` and `client_secret` \u2014 the secret is only returned once at creation time (stored as a bcrypt hash). Scopes are validated against the set of available scopes.",
        "operationId": "create_app_api_apps__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthAppCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "App created. **`client_secret` is only returned in this response.**",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAppCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid scopes."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/apps/{app_id}/status": {
      "patch": {
        "tags": [
          "apps"
        ],
        "summary": "Update app approval status (admin only)",
        "description": "Set an app's status to approved or rejected. Requires admin privileges.",
        "operationId": "update_app_status_api_apps__app_id__status_patch",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAppResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status."
          },
          "403": {
            "description": "Admin access required."
          },
          "404": {
            "description": "App not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/apps/{app_id}": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "Get an OAuth app",
        "description": "Returns a single OAuth app by its UUID. Does not include the client secret.",
        "operationId": "get_app_api_apps__app_id__get",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAppResponse"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "apps"
        ],
        "summary": "Update an OAuth app",
        "description": "Partially updates an OAuth app. Only provided fields are changed. Scopes are re-validated if provided.",
        "operationId": "update_app_api_apps__app_id__patch",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthAppUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAppResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid scopes."
          },
          "404": {
            "description": "App not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "apps"
        ],
        "summary": "Delete an OAuth app",
        "description": "Permanently deletes an OAuth app and its credentials. This cannot be undone.",
        "operationId": "delete_app_api_apps__app_id__delete",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "App deleted."
          },
          "404": {
            "description": "App not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OAuthAppCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name of the application.",
            "examples": [
              "My Cool App"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Short description of the application.",
            "examples": [
              "A demo OAuth client for testing."
            ]
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "description": "Requested OAuth scopes. Validated against available scopes: openid, profile, email, roles, date_joined, offline_access.",
            "default": [],
            "examples": [
              [
                "openid",
                "email",
                "profile"
              ]
            ]
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redirect Uris",
            "description": "Allowed OAuth callback URLs.",
            "default": [],
            "examples": [
              [
                "https://myapp.example.com/callback"
              ]
            ]
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "URL to the app icon (set via icon upload endpoint)."
          },
          "privacy_policy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Policy Url",
            "description": "URL to the app's privacy policy.",
            "examples": [
              "https://myapp.example.com/privacy"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OAuthAppCreate",
        "description": "Request body for creating a new OAuth application."
      },
      "OAuthAppCreated": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique app identifier."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "App description."
          },
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "description": "OAuth client ID (32-char hex)."
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "description": "Registered scopes."
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redirect Uris",
            "description": "Registered callback URLs."
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "App icon URL path."
          },
          "privacy_policy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Policy Url",
            "description": "Privacy policy URL."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Approval status: pending, approved, or rejected.",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "UTC creation timestamp."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "UTC last-updated timestamp."
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret",
            "description": "Client secret (returned ONCE at creation \u2014 stored as bcrypt hash, never retrievable again)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "client_id",
          "scopes",
          "redirect_uris",
          "created_at",
          "updated_at",
          "client_secret"
        ],
        "title": "OAuthAppCreated",
        "description": "OAuth application with client secret. The secret is only returned once at creation time."
      },
      "OAuthAppResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique app identifier."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "App description."
          },
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "description": "OAuth client ID (32-char hex)."
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "description": "Registered scopes."
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redirect Uris",
            "description": "Registered callback URLs."
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "App icon URL path."
          },
          "privacy_policy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Policy Url",
            "description": "Privacy policy URL."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Approval status: pending, approved, or rejected.",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "UTC creation timestamp."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "UTC last-updated timestamp."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "client_id",
          "scopes",
          "redirect_uris",
          "created_at",
          "updated_at"
        ],
        "title": "OAuthAppResponse",
        "description": "OAuth application details (without client secret)."
      },
      "OAuthAppUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "New display name."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "New description."
          },
          "scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scopes",
            "description": "New scopes (validated)."
          },
          "redirect_uris": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect Uris",
            "description": "New redirect URIs."
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "New icon URL."
          },
          "privacy_policy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Policy Url",
            "description": "New privacy policy URL."
          }
        },
        "type": "object",
        "title": "OAuthAppUpdate",
        "description": "Request body for partially updating an OAuth application. Only provided fields are changed."
      },
...
      "StatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}